- UIText 추가 중!!! - AssetModule에 LoadFont 함수 추가!! - etc

+++ Source/TGON/Text/Hash.h
@@ -14,13 +14,13 @@ namespace tgon
 {
 
 template <typename _CharType, typename std::enable_if_t<IsChar<_CharType>>* = nullptr>
-constexpr uint32_t X65599Hash(_CharType ch) noexcept
+constexpr size_t X65599Hash(_CharType ch) noexcept
 {
     return ch ^ (ch >> 16);
 }
     
 template <typename _CharType>
-constexpr uint32_t X65599Hash(const _CharType* str) noexcept
+constexpr size_t X65599Hash(const _CharType* str) noexcept
 {
     if (str == nullptr)
     {
